backend/requirements.txt
Django==2.2.16
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
gunicorn==20.0.4
Pillow==9.2.0
django_filter==21.1
python-dotenv==0.21.0
psycopg2-binary==2.8.6
drf-extra-fields==3.4.1